More refactoring and add content to README.md
