+++ src/main/java/halcyon/view/demo/DemoToolbarPanel.java
@@ -6,12 +6,13 @@ import javafx.scene.control.Button;
 import javafx.scene.layout.VBox;
 
 /**
- * Control type Toolbar window
+ * Demo Toolbar window for demonstrating. Users can make their own ToolBar window and add them {@link halcyon.HalcyonFrame}.
+ * Please, have a look {@link DemoHalcyonMain}.
  */
-public class DemoToolbarWindow extends DockNode
+public class DemoToolbarPanel extends DockNode
 {
 
-	public DemoToolbarWindow()
+	public DemoToolbarPanel()
 	{
 		super(new VBox());
 		getDockTitleBar().setVisible(false);
