+++ src/main/java/halcyon/view/demo/images/Strings.properties
@@ -1,8 +1,8 @@
-
-root.icon=/halcyon/demo/images/folder_16.png
-one.icon=/halcyon/demo/images/one_16.png
-two.icon=/halcyon/demo/images/two_16.png
-three.icon=/halcyon/demo/images/three_16.png
+# This property file is used for mapping icon to HalcyonType. Please, refer HalcyonNodeTypeExample class
+root.icon=/halcyon/view/demo/images/folder_16.png
+one.icon=/halcyon/view/demo/images/one_16.png
+two.icon=/halcyon/view/demo/images/two_16.png
+three.icon=/halcyon/view/demo/images/three_16.png
 
 # The build.date resource is automatically generated.  Leave this as the last line and do not change the key name.
 build.date=Version 1.0.0 (17.07.2015 15:42:08)
